refactor(auto-reply): extract execution test assertion helpers
The agent-runner execution test-support file crossed the 700-line
max-lines cap. Move the generic record/mock-call assertion helpers into
a sibling agent-runner-execution-assertions.test-support.ts and re-export
them, following the existing test-helpers re-export convention. Behavior
and the assertion API are unchanged; every importer keeps its current
import path.
